+++ HaveIBeenPwned/CheckTypeEnum.cs
@@ -2,8 +2,13 @@
 {
     public enum CheckTypeEnum
     {
+        [Display(Name = "Site/Service Checker")]
         SiteDomain,
+
+        [Display(Name = "Username Checker")]
         Username,
+
+        [Display(Name = "Password Checker")]
         Password
     }
 }